Add aris bash magic

+++ arviz/tests/helpers.py
@@ -586,10 +586,10 @@ def load_cached_models(eight_schools_data, draws, chains, libs=None):
         )
 
         path = os.path.join(data_directory, fname)
-        # if not os.path.exists(path):
-        with gzip.open(path, "wb") as buff:
-            _log.info("Generating and caching %s", fname)
-            pickle.dump(func(eight_schools_data, draws, chains), buff)
+        if not os.path.exists(path):
+            with gzip.open(path, "wb") as buff:
+                _log.info("Generating and caching %s", fname)
+                pickle.dump(func(eight_schools_data, draws, chains), buff)
 
         with gzip.open(path, "rb") as buff:
             _log.info("Loading %s from cache", fname)
